add Re-apply current serialization to input_constraints
